Using chart signals as extension
